+++ src/main/java/com/example/tpo/uade/Xperium/service/Direccion/DireccionServiceImpl.java
@@ -1,6 +1,5 @@
 package com.example.tpo.uade.Xperium.service.Direccion;
 
-import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,35 +17,40 @@ public class DireccionServiceImpl implements DireccionService {
 
     @Autowired
     private DireccionRepository direccionRepository;
- 
-    public Page<Direccion> getDirecciones(PageRequest pageRequest) {
-        return direccionRepository.findAll(pageRequest);
+
+    @Override
+    public Page<Direccion> getDireccionesByCompradorId(Long compradorId, PageRequest pageRequest) {
+        return direccionRepository.findByCompradorId(compradorId, pageRequest);
     }
 
-    public Optional<Direccion> getDireccionesById(Long id) {
-       return direccionRepository.findById(id);
-    }
-   
-    public Direccion createDireccion(String calle, String numero, String departamento, String codigoPostal,Comprador comprador) throws CategoriaDuplicadaException {
-        return direccionRepository.save(new Direccion(calle, numero, departamento, codigoPostal,comprador)); 
+    @Override
+    public Optional<Direccion> getDireccionesByIdAndCompradorId(Long id, Long compradorId) {
+        return direccionRepository.findByIdAndCompradorId(id, compradorId);
     }
 
-    public void deleteDireccion(Long id) {
-        direccionRepository.deleteById(id);
+    @Override
+    public Direccion createDireccion(String calle, String numero, String departamento, String codigoPostal, Comprador comprador) throws CategoriaDuplicadaException {
+        return direccionRepository.save(new Direccion(calle, numero, departamento, codigoPostal, comprador));
     }
 
-    public Direccion updateDireccion(Long id, String calle, String numero, String departamento, String codigoPostal) throws Exception {
-        Direccion direccion = direccionRepository.findById(id)
-            .orElseThrow(() -> new Exception("Direccion no encontrada"));
+    @Override
+    public void deleteDireccionByIdAndCompradorId(Long id, Long compradorId) throws Exception {
+        Optional<Direccion> direccion = direccionRepository.findByIdAndCompradorId(id, compradorId);
+        if (direccion.isPresent()) {
+            direccionRepository.deleteById(id);
+        } else {
+            throw new Exception("Dirección no encontrada o no pertenece al comprador");
+        }
+    }
+
+    @Override
+    public Direccion updateDireccion(Long id, Long compradorId, String calle, String numero, String departamento, String codigoPostal) throws Exception {
+        Direccion direccion = direccionRepository.findByIdAndCompradorId(id, compradorId)
+            .orElseThrow(() -> new Exception("Dirección no encontrada o no pertenece al comprador"));
         direccion.setCalle(calle);
         direccion.setNumero(numero);
         direccion.setDepartamento(departamento);
         direccion.setCodigoPostal(codigoPostal);
         return direccionRepository.save(direccion);
     }
-    
-    public List<Direccion> getDireccionesByCompradorId(Long compradorId) {
-    return direccionRepository.findByCompradorId(compradorId);
-    }
-    
 }
